Fix path and env (#16)
There was an issue in the path and the env that breaks the main branch

This is a fix. Basically there was a path using lowercase, but then imported as uppercase.

For the env, the quotes characters were wrong.

Also, I think is nice to provide a default API KEY for infura, users might want to change it if it get's overused, but at least it works out of the box.